makes patch aam, atm and plot_ced work
